Stuff from robin

+++ Core/Application.cpp
@@ -66,7 +66,7 @@ bool Application::Initialize(HINSTANCE hInst)
 	
 	this->gBufferShader.init(gBufferDesc);
 
-	this->mainCamera.SetProjectionMatrix(D3DX_PI/2.0f, D3DShell::self()->getAspectRatio(), 1, 1000);
+	this->mainCamera.SetProjectionMatrix((float)D3DX_PI/2.0f, D3DShell::self()->getAspectRatio(), 1, 1000);
 	this->mainCamera.SetOrthogonalMatrix(D3DShell::self()->getWidth(), D3DShell::self()->getHeight(), 1, 1000);
 	this->mainCamera.SetPosition(0.0f, 0.0f, 0.0f);
 	this->mainCamera.SetRotation(0.0f, 0.0f, 0.0f);
